requirements.txt
# pdf extraction
fitz
boto3
psycopg2
datetime
python-dotenv

# FastAPI
fastapi
uvicorn

# Dashboard
streamlit
streamlit_option_menu
pandas
plotly